Remove the crash at application ending in debug.
Necessary to make test plugin to work (otherwise tests always fail because of the crash at the end).

Due to the fact that debug tracing of smart pointer was using static variables that are invalidated before the really end of the application: some call to static functions that use these variables are done after these variables are invalidated.

Solution that is implemented here is a typical solution to avoid "static initialization order fiasco" but used here not for initialistion order but for access after a static variables have been deleted:

IMportant: notice that in statuc function, static variable are static pointer on object and not directely object. It would not work if they where not pointer.

The consequence in these object are never deleted, this is what we want! A better solution would be to rewrite entirely DebugUtil to avoig using static variable (not very good in C++)